+++ deploy/shopping-cart/templates/ingress.yaml
@@ -1,16 +1,16 @@
-# apiVersion: extensions/v1beta1
-# kind: Ingress
-# metadata:
-#   name: ingress
-#   namespace: cart
-#   annotations:
-#     kubernetes.io/ingress.class: "nginx"
-# spec:
-#   rules:
-#   - host: shopping-cart.io  
-#     http:
-#       paths:
-#       - path: /
-#         backend:
-#           serviceName: shopping-cart
-#           servicePort: 8008
+apiVersion: extensions/v1beta1
+kind: Ingress
+metadata:
+  name: ingress
+  namespace: cart
+  annotations:
+    kubernetes.io/ingress.class: "nginx"
+spec:
+  rules:
+  - host: shopping-cart.io  
+    http:
+      paths:
+      - path: /
+        backend:
+          serviceName: shopping-cart
+          servicePort: 8008
