+++ pkg/controllers/cart/cart.go
@@ -3,6 +3,7 @@ package cart
 import (
 	"encoding/json"
 	"net/http"
+	"shopping-cart/pkg/service"
 	"shopping-cart/types"
 	"shopping-cart/utils/applog"
 
@@ -11,15 +12,20 @@ import (
 
 // AddItem : handler function for PATCH /v1/cart call
 func AddItem(w http.ResponseWriter, r *http.Request) {
+	// authenticating user
 	accessToken := &types.AccessToken{}
-	if !accessToken.AuthorizeByToken(w, r) {
+	accessToken.GetTokenFromRequest(r)
+	authService := &service.AuthService{}
+	if !authService.AuthorizeByToken(w, accessToken) {
 		return
 	}
-	cartid := accessToken.GetUser().CartID
+	// proccess add to cart request
+	cartid := authService.GetUser().CartID
 	cart := &types.Cart{}
 	cart.ID = cartid
 	applog.Info("adding item to cart")
-	if cart.Validate(w, r) && cart.AddToCart(w) { 
+	cartService:=service.CartService{}
+	if cartService.Validate(w, r, cart) && cartService.AddToCart(w, cart) { 
 		w.Header().Set("Content-Type", "application/json")
 		w.WriteHeader(http.StatusOK)
 		response := map[string]interface{}{"data": cart, "status": 1}
@@ -30,15 +36,18 @@ func AddItem(w http.ResponseWriter, r *http.Request) {
 // ViewCart Get All items in cart
 func ViewCart(w http.ResponseWriter, r *http.Request) {
 	accessToken := &types.AccessToken{}
-	if !accessToken.AuthorizeByToken(w, r) {
+	accessToken.GetTokenFromRequest(r)
+	authService := &service.AuthService{}
+	if !authService.AuthorizeByToken(w, accessToken) {
 		return
-	} 
+	}
 	
-	cartid := accessToken.GetUser().CartID
+	cartid := authService.GetUser().CartID
 	cart := &types.Cart{}
 	cart.ID = cartid 
 	applog.Info("get all items from cart")
-	if cart.ViewCart(w) { 
+	cartService:=service.CartService{}
+	if cartService.ViewCart(w, cart) { 
 		w.Header().Set("Content-Type", "application/json")
 		w.WriteHeader(http.StatusOK)
 		response := map[string]interface{}{"data": cart.Items, "status": 1}
@@ -50,15 +59,17 @@ func ViewCart(w http.ResponseWriter, r *http.Request) {
 // RemoveItem : delete item from cart
 func RemoveItem(w http.ResponseWriter, r *http.Request) {
 	accessToken := &types.AccessToken{}
-	if !accessToken.AuthorizeByToken(w, r) {
+	accessToken.GetTokenFromRequest(r)
+	authService := &service.AuthService{}
+	if !authService.AuthorizeByToken(w, accessToken) {
 		return
 	}
 	params := mux.Vars(r)
-	cartid := accessToken.GetUser().CartID
+	cartid := authService.GetUser().CartID
 	cart := &types.Cart{}
     cart.ID = cartid 
-
-	if cart.RemoveItem(w, params["itemid"]) { 
+	cartService:=service.CartService{}
+	if cartService.RemoveItem(w, cart, params["itemid"]) { 
 		w.Header().Set("Content-Type", "application/json")
 		w.WriteHeader(http.StatusOK)
 		response := map[string]interface{}{"data": cart, "message": "Item Deleted Successfully", "status": 1}
@@ -68,14 +79,16 @@ func RemoveItem(w http.ResponseWriter, r *http.Request) {
 // ClearCart : delete item from cart
 func ClearCart(w http.ResponseWriter, r *http.Request) {
 	accessToken := &types.AccessToken{}
-	if !accessToken.AuthorizeByToken(w, r) {
+	accessToken.GetTokenFromRequest(r)
+	authService := &service.AuthService{}
+	if !authService.AuthorizeByToken(w, accessToken) {
 		return
-	} 
-	cartid := accessToken.GetUser().CartID
+	}
+	cartid := authService.GetUser().CartID
 	cart := &types.Cart{}
     cart.ID = cartid 
-
-	if cart.ClearCart(w) { 
+	cartService:=service.CartService{}
+	if cartService.ClearCart(w,cart) { 
 		w.Header().Set("Content-Type", "application/json")
 		w.WriteHeader(http.StatusOK)
 		response := map[string]interface{}{"data": cart, "message": "Deleted Successfully", "status": 1}
