+++ pkg/controllers/user/user.go
@@ -3,58 +3,23 @@ package user
 import (
 	"encoding/json"
 	"net/http"
-	"shopping-cart/pkg/database"
+	"shopping-cart/pkg/service"
 	"shopping-cart/types"
 	"shopping-cart/utils/applog"
-
-	"gopkg.in/mgo.v2/bson"
-
-	"github.com/jameskeane/bcrypt"
 )
 
 // RegisterUser : Register user account
 func RegisterUser(w http.ResponseWriter, r *http.Request) {
 	user := &types.User{}
 	applog.Info("Register new user")
-	if !user.Validate(w, r) {
-		return
-	}
-	db := database.Db
-	userDb := db.C("user")
-	cart := &types.Cart{}
-	cartDb:= db.C("cart")
-	// Insert
-	cart.ID = bson.NewObjectId()
-	cart.Items= []types.CartItem{}
-	cartErr := cartDb.Insert(&cart)
-	if cartErr != nil {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusInternalServerError)
-		response := map[string]interface{}{"errors": cartErr.Error(), "status": 0}
-		json.NewEncoder(w).Encode(response)
-
-	}
-	user.ID = bson.NewObjectId()
-	user.CartID = cart.ID
-	salt, _ := bcrypt.Salt(10)
-	user.Password, _ = bcrypt.Hash(user.Password, salt)
-
-	insertionErrors := userDb.Insert(&user)
-
-	if insertionErrors != nil {
-		applog.Error("error occured while registration of new user")
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusInternalServerError)
-		response := map[string]interface{}{"errors": insertionErrors.Error(), "status": 0}
-		json.NewEncoder(w).Encode(response)
-
-	} else {
+	userService := service.UserService{} 
+	if userService.Validate(w, r, user) && userService.RegisterUser(w, r, user){
 		applog.Debugf("User '%s' created successfully",user.ID)
 		w.Header().Set("Content-Type", "application/json")
 		w.WriteHeader(http.StatusOK)
 		response := map[string]interface{}{"data": user, "status": 1}
 		json.NewEncoder(w).Encode(response)
-	}
+	} 
 	applog.Info("Register of new user completed")
 }
 
@@ -62,11 +27,13 @@ func RegisterUser(w http.ResponseWriter, r *http.Request) {
 func Login(w http.ResponseWriter, r *http.Request) {
 	auth := &types.AccessTokenRequest{}
 	applog.Info("login in user")
-	if !auth.Validate(w, r) {
+	authService:= service.AuthService{}
+	if !authService.Validate(w, r, auth) {
 		applog.Debug("unable to allow login")
 		return
 	}
-	accesstoken := auth.GenerateAccessToken(w)
+	applog.Info("generating token for user")
+	accesstoken := authService.GenerateAccessToken(w, auth)
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
 	response := map[string]interface{}{"data": map[string]interface{}{
@@ -79,12 +46,15 @@ func Login(w http.ResponseWriter, r *http.Request) {
 
 // LogOut : logout from account
 func LogOut(w http.ResponseWriter, r *http.Request) {
+	// authenticating user
 	accessToken := &types.AccessToken{}
-	if !accessToken.AuthorizeByToken(w, r) {
+	accessToken.GetTokenFromRequest(r)
+	authService := &service.AuthService{}
+	if !authService.AuthorizeByToken(w, accessToken) {
 		return
-	}
-	user := accessToken.GetUser()
-	accessToken.Remove()
+	} 
+	user := authService.GetUser()
+	authService.Remove(accessToken)
 
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
